[refactor] preparation for PR adressing #638
